requirements.txt
pandas==2.0.0
numpy==1.24.3
matplotlib==3.7.1
pyyaml==6.0
pytest==7.3.1
alpaca-trade-api==3.0.2
ib_insync==0.9.85
scikit-learn==1.2.2
statsmodels==0.14.0
jupyter==1.0.0
pyarrow==12.0.0
requests==2.30.0
tqdm==4.65.0